# frontend/src/lib/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "InvoiceFlow API",
    "description": "E-invoice processing platform with AI extraction and global compliance (PEPPOL, ZATCA, Brazil NFe, India IRP, Mexico CFDI, CTC reporting, and post-audit archival).",
    "version": "0.2.0"
  },
  "servers": [
    { "url": "http://localhost:8000", "description": "Local development" },
    { "url": "/api/v1", "description": "Production (Nginx proxy)" }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from POST /api/v1/auth/login"
      }
    },
    "schemas": {
...
      "TokenRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" }
        },
        "required": ["email", "password"]
      },
      "TokenResponse": {
        "type": "object",
        "properties": {
          "access_token": { "type": "string" },
          "token_type": { "type": "string", "default": "bearer" },
          "user_id": { "type": "string" },
          "tenant_id": { "type": "string" }
        },
        "required": ["access_token", "user_id", "tenant_id"]
      },
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" },
          "full_name": { "type": "string", "nullable": true },
          "tenant_name": { "type": "string" },
          "tenant_slug": { "type": "string" },
          "country_code": { "type": "string", "default": "US" }
        },
        "required": ["email", "password", "tenant_name", "tenant_slug"]
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string" },
          "full_name": { "type": "string", "nullable": true },
          "role": { "type": "string" },
          "tenant_id": { "type": "string", "format": "uuid" }
        },
        "required": ["id", "email", "role", "tenant_id"]
      },
...
      "DashboardStats": {
        "type": "object",
        "properties": {
          "invoices_today": { "type": "integer" },
          "success_rate": { "type": "number" },
          "pending": { "type": "integer" },
          "total_processed": { "type": "integer" }
        },
        "required": ["invoices_today", "success_rate", "pending", "total_processed"]
      },
      "ChartData": {
        "type": "object",
        "properties": {
          "labels": { "type": "array", "items": { "type": "string" } },
          "data": { "type": "array", "items": { "type": "integer" } }
        },
        "required": ["labels", "data"]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "detail": { "type": "string" }
        },
        "required": ["detail"]
      }
    }
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": ["auth"],
        "summary": "Login and receive JWT token",
        "operationId": "login",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TokenRequest" } } } },
        "responses": {
          "200": { "description": "Login successful", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TokenResponse" } } } },
          "401": { "description": "Invalid credentials", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "tags": ["auth"],
        "summary": "Register a new user and tenant",
        "operationId": "register",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RegisterRequest" } } } },
        "responses": {
          "201": { "description": "Registration successful", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserResponse" } } } },
          "409": { "description": "Email already registered", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": ["auth"],
        "summary": "Get current user info",
        "operationId": "getCurrentUser",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Current user info", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserResponse" } } } },
          "401": { "description": "Not authenticated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
...
    "/api/v1/analytics/dashboard": {
      "get": {
        "tags": ["analytics"],
        "summary": "Get dashboard statistics",
        "operationId": "getDashboardStats",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Dashboard statistics", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DashboardStats" } } } }
        }
      }
    },
    "/api/v1/analytics/charts/daily": {
      "get": {
        "tags": ["analytics"],
        "summary": "Get daily invoice volume for charts",
        "operationId": "getDailyVolume",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "days", "in": "query", "schema": { "type": "integer", "default": 7 } }],
        "responses": {
          "200": { "description": "Daily volume data", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ChartData" } } } }
        }
      }
    },
    "/api/v1/analytics/charts/by-country": {
      "get": {
        "tags": ["analytics"],
        "summary": "Get invoice counts by country",
        "operationId": "getByCountry",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Country distribution data", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ChartData" } } } }
        }
      }
    },
...
  },
  "security": [{ "BearerAuth": [] }]
}